+++ packages/components/CHANGELOG.md
@@ -87,7 +87,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 - ✅Multi Select
 - ✅Optgroup
 - ✅Pin Code
-- ⌛Radio Group
+- ✅Radio Group
 - ✅Segmented Control
 - ✅Select
 - ✅Switch